VideoTableWithSearch component
